% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggslice.R
\name{ggslice}
\alias{ggslice}
\title{Title}
\usage{
ggslice(
  data,
  z,
  z_q_range = 0.5,
  z_value = NULL,
  tolerance = 0.05,
  tol_value = NULL
)
}
\arguments{
\item{data}{}

\item{z}{}

\item{z_q_range}{}

\item{z_value}{}

\item{tolerance}{}

\item{tol_value}{}
}
\description{
Title
}
